feat: 돈 관련 BigDecimal로 타입 수정

+++ src/main/java/lotto/domain/Rank.java
@@ -1,20 +1,21 @@
 package lotto.domain;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.function.BiFunction;
 
 public enum Rank {
-    FIRST_PLACE(2000000000,  (val, isBonus) -> val == 6),
-    SEC0ND_PLACE(30000000, (val, isBonus) -> val == 5 && isBonus),
-    THIRD_PLACE(1500000, (val, isBonus) -> val == 5 && !isBonus),
-    FOURTH_PLACE(50000, (val, isBonus) -> val == 4),
-    FIFTH_PLACE(5000, (val, isBonus) -> val == 3),
-    UNRANKED(0, (val, isBonus) -> val < 3);
+    FIRST_PLACE(BigDecimal.valueOf(2000000000),  (val, isBonus) -> val == 6),
+    SEC0ND_PLACE(BigDecimal.valueOf(30000000), (val, isBonus) -> val == 5 && isBonus),
+    THIRD_PLACE(BigDecimal.valueOf(1500000), (val, isBonus) -> val == 5 && !isBonus),
+    FOURTH_PLACE(BigDecimal.valueOf(50000), (val, isBonus) -> val == 4),
+    FIFTH_PLACE(BigDecimal.valueOf(5000), (val, isBonus) -> val == 3),
+    UNRANKED(BigDecimal.valueOf(0), (val, isBonus) -> val < 3);
 
     private final BiFunction<Integer, Boolean, Boolean> expression;
-    private final int prize;
+    private final BigDecimal prize;
 
-    Rank(int prize, BiFunction<Integer, Boolean, Boolean> expression) {
+    Rank(BigDecimal prize, BiFunction<Integer, Boolean, Boolean> expression) {
         this.prize = prize;
         this.expression = expression;
     }
@@ -26,7 +27,7 @@ public enum Rank {
             .orElse(UNRANKED);
     }
 
-    public int getPrize() {
+    public BigDecimal getPrize() {
         return this.prize;
     }
